Struct directx::dxgi::Adapter

pub struct Adapter {
    // some fields omitted
}

Methods

impl Adapter

unsafe fn from_raw(raw: ComPtr<IDXGIAdapter>) -> Adapter

fn outputs(&self) -> OutputIter

Returns an iterator over all available outputs.

fn check_interface_support(&self, iid: REFGUID) -> Option<LARGE_INTEGER>

Check to see if an interface is supported on this adapter.

If the interface is supported, Ok is returned, containing the supported version. If the interface is not supported, Err is returned.

fn description(&self) -> String

fn vendor_id(&self) -> UINT

fn device_id(&self) -> UINT

fn subsystem_id(&self) -> UINT

fn revision(&self) -> UINT

fn dedicated_video_memory(&self) -> SIZE_T

fn dedicated_system_memory(&self) -> SIZE_T

fn shared_system_memory(&self) -> SIZE_T

fn adapter_luid(&self) -> LUID